Temporary fix
-Removed Bank 21 as it is no longer needed.
---
Increasing the size of the KeyboardData_Player variable crashed the autodeck machines. Reverted the KeyboardData_Player variable to its defaults. Until bank 6 is completely disassembled I'm gonna leave the player keyboard for what it is and continue translating the rest with the save games I've already prepared. I've added a backup of my version of the KeyboardData_Player variable in a .txt file.